refactor(frontend): extract RootLayout shell from App.tsx into navigation/RootLayout
Move the 1750-line RootLayout component together with ProjectMenu,
GlobalMenu, the compact-topbar icon helper and their constants verbatim
into src/navigation/RootLayout.tsx. App.tsx now only contains the lazy
route imports, redirect helpers, error boundaries, router factory and
the App component (~330 lines). No behavior changes.
